+++ tests/array/array/testArray1RefsBase.cpp
@@ -2,7 +2,7 @@ TEST_F(ARRAY_DIMS_SCALAR_TEST_REFS, when_resize_then_sizeChanges)
 {
   EXPECT_EQ(ARRAY_SIZE, arr.size());
 
-  uint newLen = ARRAY_SIZE + 1;
+  size_t newLen = ARRAY_SIZE + 1;
   arr.resize(newLen);
 
   EXPECT_EQ(newLen, arr.size());
@@ -10,7 +10,7 @@ TEST_F(ARRAY_DIMS_SCALAR_TEST_REFS, when_resize_then_sizeChanges)
 
 TEST_F(ARRAY_DIMS_SCALAR_TEST_REFS, when_getIndexWithParentheses_then_refReturned)
 {
-  uint i = 1;
+  size_t i = 1;
   arr(i) = VAL;
 
   EXPECT_EQ(VAL, arr(i));
@@ -18,7 +18,7 @@ TEST_F(ARRAY_DIMS_SCALAR_TEST_REFS, when_getIndexWithParentheses_then_refReturne
 
 TEST_F(ARRAY_DIMS_SCALAR_TEST_REFS, given_constCompressedArray_when_getIndexWithBrackets_then_valReturned)
 {
-  uint i = 1;
+  size_t i = 1;
   arr[i] = VAL;
 
   const array1<SCALAR> arrConst = arr;
@@ -28,7 +28,7 @@ TEST_F(ARRAY_DIMS_SCALAR_TEST_REFS, given_constCompressedArray_when_getIndexWith
 
 TEST_F(ARRAY_DIMS_SCALAR_TEST_REFS, given_constCompressedArray_when_getIndexWithParentheses_then_valReturned)
 {
-  uint i = 1;
+  size_t i = 1;
   arr[i] = VAL;
 
   const array1<SCALAR> arrConst = arr;
