+++ tests/array/array/testArray2Base.cpp
@@ -15,12 +15,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromConstView_then_sizeConse
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -35,12 +35,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromConstView_then_performsD
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -48,8 +48,8 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromConstView_then_performsD
   ZFP_ARRAY_TYPE arr2(v);
 
   /* verify array entries */
-  for (uint j = 0; j < viewLenY; j++) {
-    for (uint i = 0; i < viewLenX; i++) {
+  for (size_t j = 0; j < viewLenY; j++) {
+    for (size_t i = 0; i < viewLenX; i++) {
       EXPECT_EQ(arr(offsetX + i, offsetY + j), arr2(i, j));
     }
   }
@@ -74,12 +74,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromView_then_sizeConserved)
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -94,12 +94,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromView_then_performsDeepCo
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -107,8 +107,8 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromView_then_performsDeepCo
   ZFP_ARRAY_TYPE arr2(v);
 
   /* verify array entries */
-  for (uint j = 0; j < viewLenY; j++) {
-    for (uint i = 0; i < viewLenX; i++) {
+  for (size_t j = 0; j < viewLenY; j++) {
+    for (size_t i = 0; i < viewLenX; i++) {
       EXPECT_EQ(arr(offsetX + i, offsetY + j), arr2(i, j));
     }
   }
@@ -133,12 +133,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromFlatView_then_sizeConser
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -153,12 +153,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromFlatView_then_performsDe
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -166,8 +166,8 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromFlatView_then_performsDe
   ZFP_ARRAY_TYPE arr2(v);
 
   /* verify array entries */
-  for (uint j = 0; j < viewLenY; j++) {
-    for (uint i = 0; i < viewLenX; i++) {
+  for (size_t j = 0; j < viewLenY; j++) {
+    for (size_t i = 0; i < viewLenX; i++) {
       EXPECT_EQ(arr(offsetX + i, offsetY + j), arr2(i, j));
     }
   }
@@ -193,12 +193,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromNestedView_then_sizeCons
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -213,12 +213,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromNestedView_then_performs
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -226,8 +226,8 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromNestedView_then_performs
   ZFP_ARRAY_TYPE arr2(v);
 
   /* verify array entries */
-  for (uint j = 0; j < viewLenY; j++) {
-    for (uint i = 0; i < viewLenX; i++) {
+  for (size_t j = 0; j < viewLenY; j++) {
+    for (size_t i = 0; i < viewLenX; i++) {
       EXPECT_EQ(arr(offsetX + i, offsetY + j), arr2(i, j));
     }
   }
@@ -255,12 +255,12 @@ TEST_P(TEST_FIXTURE, when_construct1dCompressedArrayFromNestedView1_then_sizeCon
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -276,23 +276,23 @@ TEST_P(TEST_FIXTURE, when_construct1dCompressedArrayFromNestedView1_then_perform
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
   ZFP_ARRAY_TYPE::nested_view v(&arr, offsetX, offsetY, viewLenX, viewLenY);
-  uint y = 1;
+  size_t y = 1;
   ZFP_ARRAY_TYPE::nested_view1 v2 = v[y];
 
   array1<SCALAR> arr2(v2);
 
   /* verify array entries */
-  for (uint i = 0; i < viewLenX; i++) {
+  for (size_t i = 0; i < viewLenX; i++) {
     EXPECT_EQ(arr(offsetX + i, offsetY + y), arr2(i));
   }
 
@@ -317,12 +317,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromPrivateConstView_then_si
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -337,12 +337,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromPrivateConstView_then_pe
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -350,8 +350,8 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromPrivateConstView_then_pe
   ZFP_ARRAY_TYPE arr2(v);
 
   /* verify array entries */
-  for (uint j = 0; j < viewLenY; j++) {
-    for (uint i = 0; i < viewLenX; i++) {
+  for (size_t j = 0; j < viewLenY; j++) {
+    for (size_t i = 0; i < viewLenX; i++) {
       EXPECT_EQ(arr(offsetX + i, offsetY + j), arr2(i, j));
     }
   }
@@ -376,12 +376,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromPrivateView_then_sizeCon
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -396,12 +396,12 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromPrivateView_then_perform
 {
   ZFP_ARRAY_TYPE arr(inputDataSideLen, inputDataSideLen, getRate(), inputDataArr);
 
-  uint offsetX = 5;
-  uint viewLenX = 3;
+  size_t offsetX = 5;
+  size_t viewLenX = 3;
   EXPECT_LT(offsetX + viewLenX, arr.size_x());
 
-  uint offsetY = 1;
-  uint viewLenY = 3;
+  size_t offsetY = 1;
+  size_t viewLenY = 3;
   EXPECT_LT(offsetY + viewLenY, arr.size_y());
 
   /* create view and construct from it */
@@ -409,8 +409,8 @@ TEST_P(TEST_FIXTURE, when_construct2dCompressedArrayFromPrivateView_then_perform
   ZFP_ARRAY_TYPE arr2(v);
 
   /* verify array entries */
-  for (uint j = 0; j < viewLenY; j++) {
-    for (uint i = 0; i < viewLenX; i++) {
+  for (size_t j = 0; j < viewLenY; j++) {
+    for (size_t i = 0; i < viewLenX; i++) {
       EXPECT_EQ(arr(offsetX + i, offsetY + j), arr2(i, j));
     }
   }
